refactor: Extract thread recomputation methods out of Thread and into their own file
